erp5_crm: Display stop date to events that have duration
stop date is very useful to Phone Call and Visit. For example, if you want to calculate the time spent in one phone call, you can handle it using start and stop date.

But, Event_view does not display stop_date. So, we display stop_date if hasDuration is True.