Products.CMFActivity.ActivityTool: Avoid activity execution-time user lookup.
When spawning an activity, store the security context user in the Message
itself, so the same user can be re-created during activity execution
without having to look the user up, then looking up their assignments
(and other documents involved in group computation), etc, saving the cost
of these calls.
Also, this follows the principle that
  foo.activate(...).bar(...)
should be equivalent to its "immediate execution" version
  foo.bar(...)
by ensuring that the security context of the activity is the same as the
one which was applied to the code which spawned that activity,
independently of any intermediate configuration change - hence improving
(deferred and fragmentary) transaction isolation.
Also, remove redundant user_name argument of Message.changeUser method.