runUnitTest: set HTTPRequest.retry_max_count
In case of ConflictError, request will be retied for
HTTPRequest.retry_max_count times. For Zope 2, the default value of the
class attribute was 3, but on Zope 4, default value is 0 and this is set
to 3 on startup (by Zope2.Startup.handlers.root_wsgi_handler).
This startup code is not executed by runUnitTest, we need to set
HTTPRequest.retry_max_count somewhere, otherwise ConflictError are not
retried during tests.