ui_test_core: replace form to update reference snapshot by a download link
In case of snapshot test failure, we generated a form in a data:test/html
link. This was working on chrome at the time, but it no longer work these
days, probably because we switched to Same-Site cookie for authentication.
Anyway, this also had the problem that it's not so convenient to use
when updating the reference from the actual screenshot took from the
test running on test node.

These change this to generate a simple download link, and the workflow to
update reference snapshot becomes to click this link to get the new image,
upload it to development zope instance and re-export the business template.