CodingStyleTest: do not install erp5_upgrader test dependencies
We were installing erp5_upgrader with its test dependencies, so it was
installing erp5_full_text_mroonga_catalog. We were also installing
erp5_full_text_myisam_catalog, so we ended up with two full text
business template installed (and the order of this installation was not
deterministic).
This was also problematic, because we were installing all the test
dependencies of erp5_upgrader, but we don't really need them here and
one goal of CodingStyleTest is to install only the dependencies of
tested business templates, as a way to verify that dependencies are
correctly specified.

This new approach hooks at another level to install erp5_upgrader
without its test dependencies - and if a full text business template is
needed, it should be listed as depencencies.

This also reverts commit ab44309fd0259b717d4c10619216d73fa3d64328 (
full_text_myisam_catalog: skip coding style test) that is no longer
needed with this new approach.