BusinessTemplate: do not export PythonScripts' Script_magic
This is used to trigger recompilation of bytecode, but we don't
export the byte code so we don't need to export it.
This also prevent diffs when the Script_magic changed, like it's the
case with Products.PythonScripts 4.2 , with commit 590125a (Force
recompilation of scripts., 2017-10-23). This commit is also
backported in ERP5 SlapOS software release.