py3: BusinessTemplate: Fix reading of BusinessTemplate files.
With Python2, properties were all str(). with Python3, we need to distinguish
binary-like properties from text-like properties.

This should ideally be implemented by checking PropertySheet elementary_type
(such as `data` for bytes() and `string` or `text` for str()) for each property.
For now (bootstrap/addERP5Site) though, let's consider all properties in .xml to
be UTF-8 str() and statically define which is which for non-.xml files.