AlarmTool: handle automatic solve with alarms owned by system user
Business templates are installed by system user, which is a special
user which is not returned by getWrappedOwner. Because of this, the
"fixing problems or activating a disabled alarm is not allowed" was
raised when checking if the owner of the alarm has manage portal
permission on the alarm.

This switches the implementation to explicit creation of the user
when user id is the system user, so that we have a real user that
has the permission.