Zelenium: catch potential errors when making (HTML) link for failed test
When a test fail, we make a data-url link with the HTML of the current
page, so that we can easily investigate test failures n test nodes.
We should not let errors that might happen here propagate, otherwise
the test result is not created and the test runner does not detect
that the test is finished.

One case that caused such errors was failed assertion just after
using goBack command without waiting