composition: only consider validated when looking up effective models
composition API, which is used to select the "most applicable" model
(Trade Condition, Transformation, Pay Sheet Model, etc) with the same
reference was only discarding deleted and validated models, but it
was selecting draft models.

This leads to this kind of problem:
 - A validated model exist, with an effective date
 - A new model is made by cloning the validated one, edited to set
  effective date from now and validated
 - The first validated model is also cloned by mistake, but not
  validated
In such case, the model cloned by mistake might be used, which from
user point of view is wrong, because user validated a model, so if
the system use a draft instead, that's a very wrong behaviour.

The first version of the patch was to simply consider validated models,
but it had a big impact on the test suite, so the second version of
the patch will accept a draft model if it is referenced directly
(ie. order -specialise-> model ), but not during the lookup of
another model to replace a model from the chain.