TimerServer: use task_dispatcher.add_task instead of high level add_task
create_server returns either a BaseWSGIServer or a MultiSocketServer,
but add_task only exists in BaseWSGIServer - and the implementation is
simply [1]:

    def add_task(self, task):
        self.task_dispatcher.add_task(task)

by using self.task_dispatcher.add_task directly, we are compatible with
both cases.

This is relevant when running timerserver with runUnitTest, where we
create multiple sockets, one for http and one for web-dav.

1: https://github.com/Pylons/waitress/blob/v1.4.4/src/waitress/server.py#L290-L291