BusinessTemplate: drop useless fixZSQLMethod in SkinTemplateItem.install
In SkinTemplateItem.install, self._objects contains entries for skin
folders and for all skins. objectValues method calls was called for skin
folders (as expected) and also for all skins, which acquire objectValues
from skin folder and do the work again.

This simplifies this by only running this for skin folders.