Listbox: refactor computation of "default" URL
"default" URL is the URL when listbox column does not use URL column
and when the brains do not have a getListItemUrl method.

Move the computation in a lazyMethod, so that it is computed only once
per line instead of once per cell.

Use make_query to build the URL,
which properly escape and fixes a problem &amp; in URL was encoded twice
( bug_module/1137 ). This also fixes the case that selection_index was
passed as a string here, it is now passed as an int, so we needed to
adjust a page template, because request/selection_index was false for
the first index which is 0.