core/xhtml_style: support request/selection_index to be an int
Most of the time, selection_index is a string, so we can test if it is
True to determine if the request parameter is passed - and in that case
include the navigation buttons to go to previous/next document of the
selection.

This change such tests to check for the truthiness of request/selection_name
instead, because when selection_index is an int, it is false with the
first document of the selection, because selection_index is 0.

This prepares for being able to use ZTUtils.make_query to build navigation
URLs including selection_name and selection_index, because make_query
also marshall argument types.