Listbox: refactor computation of "default" URL
"default" URL is the URL when listbox column does not use URL column
and when the brains do not have a getListItemUrl method.

Move the computation in a lazyMethod, so that it is computed only once
per line instead of once per cell.

Also remove a try/except, I don't think this code is supposed to get an
AttributeError