sql_browser: improve display of results with big numbers
Javascript does not support big integers, so the backend have to be
careful not to return too large numbers in the json. Nowadays there
is bigint, but not equivalent for decimal numbers, so the backend
tries to detect if numbers are too big and returns them as string
in that case - which is better than returning a different number
because it displays correctly in the result table. The pivot table
will still wrong, but at least the table of results is OK.