stripe: store more info in workflow transition
Add a comment for more traceability.

Add the payment_status, so that it can be accessed directly in an
interaction workflow on `complete`. This eliminate the needs for
activity scheduling (that was here so that interaction workflow on
complete can access the related HTTP Exchange - so it needed to wait
for the related HTTP Exchange to be indexed)