hal_json_style: fix double translation of worklist message
There was a regression with 45c03413ac (hal_json_style: prevent a
translation of worklist with document count, 2021-05-10), the title
of the worklist was translated twice. This was supposed to be
catched by the test, by checking that the mocked gettext had only
one call for "Draft to Validate", assuming that the translation of
"Draft to Validate" was "Draft to Validate" and that a double
translation would cause this to be translated twice, but because of
a side effect of another test inserting a translation for
"Draft to Validate", this problem was not noticed.

This fixes the double translation and adjust the test to not depend
on the actual content of message catalog.