tests: remove broken coverage support
coverage needs to be started before importing the modules, so it can
not be started so late in runUnitTest, it has to be started before
importing any Products/ERP5* code, ie. in slapos wrapper script.

See https://coverage.readthedocs.io/en/6.5.0/faq.html#q-why-do-the-bodies-of-functions-show-as-executed-but-the-def-lines-do-not
for more details about the rationale