ERP5TypeFunctionalTestCase: make sure activities are not subscribed
Invoking Zuite_waitForActivities when portal_activities is already
subscribed and while another thread from this process is already
executing activities causes infinite "Got a lock error, retrying...".

This situation happens when developer makes a mistake (runs the
tests without unsubscribing activities), so be nice to the developer
so that they don't have to restart Zope.