ProcessingNodeTestCase: Adjust _p_resolveConflict for more than 2 nodes
Usually we run tests with two nodes, so when there's a conflict it's
always between the "main" node and the processing node, so we always
have a distributing node during conflict.

But when running with, for example, 2 nodes (the main node + 2 nodes),
the conflict might occur between the two processing nodes and after
the resolution we still don't have a distributing node.