gadget_html_viewer: only set a border for tables with border=1 attribute
gadget_html_viewer had styles to add a border for tables, so that they
look similar to tables added in CKEditor. One problem is that all tables
had borders, because HTML emails are often using tables to position the
content, when viewing an Mail Message from ERP5 back office interface
we could see all the borders, which looked a bit ugly.

These changes bring back a behavior closer to default style sheet of
browsers: tables have no border by default, but a rule is added to
emulate the rendering of border="1" attributes, so that tables edited
with CKEditor have borders and look the same in the editor or in the
html viewer.