accounting: keep "??? (value)" to display invalid bank account
The changes from ae15e7e136 (accounting: show the context when
showing invalid bank accounts, 2022-12-21) were a regression because
with the previous behaviour, by just looking at the transactions,
users could see with the ??? that something was wrong with the bank
account. After these changes they had to open the select element.

This restore somehow the previous behaviour by displaying the label
with ??? marker, but since we have the label, we display the
label (bank account reference) instead of the value (the relative URL)